rename cogmanagermixin to cogmanagermixin
